--- src/main/java/Example.java
@@ -1,19 +1,0 @@
-import org.springframework.boot.*;
-import org.springframework.boot.autoconfigure.*;
-import org.springframework.stereotype.*;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@EnableAutoConfiguration
-public class Example {
-
-    @RequestMapping("/")
-    String home() {
-        return "Hello World - v3!";
-    }
-
-    public static void main(String[] args) throws Exception {
-        SpringApplication.run(Example.class, args);
-    }
-
-}
