+++ src/main/java/com/example/myproject/Example.java
@@ -1,0 +1,20 @@
+/* (C) Private company */
+package com.example.myproject;
+
+import org.springframework.boot.*;
+import org.springframework.boot.autoconfigure.*;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@EnableAutoConfiguration
+public class Example {
+
+  @RequestMapping("/")
+  String home() {
+    return "Hello World - v3!";
+  }
+
+  public static void main(String[] args) throws Exception {
+    SpringApplication.run(Example.class, args);
+  }
+}
